Fix: 전문가 Progress 저장 방식 수정

See merge request s12-final/S12P31S108!136

+++ BE/app/schemas/message.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, Field
 
 class Progress(BaseModel):
-    current: int = Field(..., description="현재 진행 중인 특허 순번")
+    current: float = Field(..., description="현재 진행 중인 특허 순번")
     total: int = Field(..., description="전체 특허 개수")
     percentage: float = Field(..., description="진행률(%)")
